docs/source/requirements.txt
sphinx>=1.8.3
sphinx_rtd_theme
sphinxext.opengraph
sphinx_tabs
sphinx_prompt
sphinxcontrib.images